Split out OutputChannelTransport and WinstonToolkitLogger
