+++ src/shared/logger/logger.ts
@@ -8,14 +8,13 @@ import * as os from 'os'
 import * as path from 'path'
 import * as vscode from 'vscode'
 import * as nls from 'vscode-nls'
-import * as winston from 'winston'
-import * as Transport from 'winston-transport'
 import { extensionSettingsPrefix } from '../constants'
 import { mkdir } from '../filesystem'
 import { fileExists } from '../filesystemUtilities'
 import { DefaultSettingsConfiguration, SettingsConfiguration } from '../settingsConfiguration'
 import { registerCommand } from '../telemetry/telemetryUtils'
 import { Loggable } from './loggableType'
+import { WinstonToolkitLogger } from './winstonToolkitLogger'
 
 const localize = nls.loadMessageBundle()
 
@@ -152,120 +151,5 @@ function getDefaultLogPath(): string {
         return path.join(os.homedir(), 'Library', 'Application Support', LOG_RELATIVE_PATH, DEFAULT_LOG_NAME)
     } else {
         return path.join(os.homedir(), '.config', LOG_RELATIVE_PATH, DEFAULT_LOG_NAME)
-    }
-}
-
-function formatMessage(level: LogLevel, message: Loggable[]): string {
-    // TODO : Look into winston custom formats - https://github.com/winstonjs/winston#creating-custom-formats
-    let final: string = `${makeLogTimestamp()} [${level.toUpperCase()}]:`
-    for (const chunk of message) {
-        if (chunk instanceof Error) {
-            final = `${final} ${chunk.stack}`
-        } else {
-            final = `${final} ${chunk}`
-        }
-    }
-
-    return final
-}
-
-function makeLogTimestamp(): string {
-    return moment().format('YYYY-MM-DD HH:mm:ss')
-}
-
-export class WinstonToolkitLogger implements Logger, vscode.Disposable {
-    // forces winston to output only pre-formatted message
-    private static readonly LOG_FORMAT = winston.format.printf(({ message }) => {
-        return message
-    })
-
-    public readonly level: LogLevel
-
-    private readonly logger: winston.Logger
-    private disposed: boolean = false
-
-    public constructor(logLevel: LogLevel) {
-        this.logger = winston.createLogger({
-            format: winston.format.combine(WinstonToolkitLogger.LOG_FORMAT),
-            level: logLevel
-        })
-
-        this.level = logLevel
-    }
-
-    public logToFile(logPath: string): void {
-        this.logger.add(new winston.transports.File({ filename: logPath }))
-    }
-
-    public logToOutputChannel(outputChannel: vscode.OutputChannel): void {
-        this.logger.add(
-            new OutputChannelTransport({
-                outputChannel
-            })
-        )
-    }
-
-    public debug(...message: Loggable[]): void {
-        this.writeToLogs(message, 'debug')
-    }
-
-    public verbose(...message: Loggable[]): void {
-        this.writeToLogs(message, 'verbose')
-    }
-
-    public info(...message: Loggable[]): void {
-        this.writeToLogs(message, 'info')
-    }
-
-    public warn(...message: Loggable[]): void {
-        this.writeToLogs(message, 'warn')
-    }
-
-    public error(...message: Loggable[]): void {
-        this.writeToLogs(message, 'error')
-    }
-
-    public dispose() {
-        if (!this.disposed) {
-            this.logger.close()
-            this.logger.clear()
-            this.disposed = true
-        }
-    }
-
-    private writeToLogs(message: Loggable[], level: LogLevel): void {
-        if (this.disposed) {
-            throw new Error('Cannot write to disposed logger')
-        }
-
-        const formattedMessage = formatMessage(level, message)
-        this.logger.log(level, formattedMessage)
-    }
-}
-
-interface LogEntry {
-    level: string
-    message: string
-}
-
-export class OutputChannelTransport extends Transport {
-    private readonly outputChannel: vscode.OutputChannel
-
-    public constructor(
-        options: Transport.TransportStreamOptions & {
-            outputChannel: vscode.OutputChannel
-        }
-    ) {
-        super(options)
-
-        this.outputChannel = options.outputChannel
-    }
-
-    public log(info: LogEntry, next: () => void): void {
-        setImmediate(() => {
-            this.outputChannel.appendLine(info.message)
-        })
-
-        next()
     }
 }
